+++ Magic/CHANGELOG.md
@@ -16,6 +16,9 @@
  - Add `custom_model_data` mage attribute, reads CMD from the item in the main hand
  - Support variables in the `/mage configure` command, e.g. `/mage configure @s mana mana_max - 10`
    - Note that you could already do `/mage configure @s mana_max x + 50`, I just forgot that existed.
+ - Add ValhallaMMO integration
+   - Levels exposed as attributes, one attribute per profile type.
+     - Format is `valhalla_level_(profile)`, e.g. `valhalla_level_archery`
 
 # 10.10.7